Cmake: add gnu source flag in order to get strptime and glob gnu specific flags available. Tun on optimization for chunk and meta server RPC parser modules as a workaround of cygwin compile failure due too many sections to make compiler inline functions. Tools: fix function calls argument types that must be chunkOff_t instead of platform dependent off_t. Add required cast to max() invocation in order to deal with platform specific off_t in stat structure. C bindings: add missing include file required for timespec definition. Meta server: eliminate compiler signed / unsigned compiler warning. Test scripts: add sleep between copy test completion and test directory rename on cygwin in order to ensure that all the files in this directory are closed by cygwin runtime / wrappers prior to mv invocation. Fix extra erasure test invocation condition.

+++ src/cc/meta/MetaVrSM.cc
@@ -684,7 +684,7 @@ private:
             RetryStartViewChange("not sufficient number of noded responded");
             return;
         }
-        if (theSz < mActiveCount &&
+        if (theSz < (size_t)mActiveCount &&
                 mStartViewCompletionCount < mActiveCount &&
                 TimeNow() <=
                     mViewChangeStartTime + mConfig.GetPrimaryTimeout()) {
